Save cookies in the credentials file, not passwords.
I'm not willing to put my password in a plain text file, so I changed how the
credentials work. The save_credentials.sh script will prompt for your gwtproject.org
password, authenticate, and store the cookie. (It doesn't work with OTP.)

The cookie expires after a day, so you'll have to re-run the script to update the
credentials file. The expiration time is an App Engine setting, so we could change
it to a week if that's terribly inconvenient.

After saving your credentials, run 'upload.sh credentials' to deploy to production.
You can also run 'upload.sh localhost' to deploy to a local dev appengine running
on port 8080. Deploying locally doesn't use your credentials file, so you don't
have to edit it.

Also removed the Credentials class since it seemed redundant after this change.

Change-Id: I6caa5bd1572fdd2df69725fea22c7ec7c61abd5c